Added C4263 warning to windows builds to flush out virtual function prototype bugs in derived classes. Had to rename a bunch of stuff to avoid name clashes.

+++ LgiIde/Code/GDebugContext.cpp
@@ -769,7 +769,7 @@ void GDebugContext::OnFileLine(const char *File, int Line, bool CurrentIp)
 	}
 }
 
-int GDebugContext::Write(const void *Ptr, int Size, int Flags)
+ssize_t GDebugContext::Write(const void *Ptr, ssize_t Size, int Flags)
 {
 	if (DebuggerLog && Ptr)
 	{
